+++ packages/ui/src/components/organisms/Modal/index.ts
@@ -1,0 +1,8 @@
+export { Modal } from './Modal.native'
+export type {
+  IModalProps,
+  IModalHeaderProps,
+  IModalBodyProps,
+  IModalFooterProps,
+  IModalFooterAction,
+} from './Modal.tsx'